requirements.txt
quart>=0.18.0
httpx>=0.24.0
python-dotenv>=0.19.0
hypercorn>=0.14.0